--- packages/ra-dependent-input/package.json
@@ -1,19 +1,0 @@
-{
-    "name": "ra-dependent-input",
-    "version": "2.0.0-beta3",
-    "description":
-        "A component for displaying input depending on other inputs values in react-admin",
-    "main": "./lib/index.js",
-    "repository": "https://github.com/marmelab/react-admin.git",
-    "authors": ["Gildas Garcia"],
-    "license": "MIT",
-    "dependencies": {
-        "lodash": "~4.17.5",
-        "prop-types": "~15.6.0"
-    },
-    "peerDependencies": {
-        "react": "^15.6.2",
-        "react-admin": "^2.0.0-beta1",
-        "react-dom": "^15.6.2"
-    }
-}
