--- packages/ra-dependent-input/src/index.js
@@ -1,2 +1,0 @@
-export DependentInput from './DependentInput';
-export DependentField from './DependentField';
