--- packages/ra-dependent-input/src/setupTests.js
@@ -1,4 +1,0 @@
-import Enzyme from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-Enzyme.configure({ adapter: new Adapter() });
